refactor(fleet): extract general instructions to fleet-instructions.ts (P4.1)
Move GENERAL_INSTRUCTIONS (the coordinator's system prompt, ~134 lines),
INSTRUCTIONS_FILENAME (per-backend file mapping), and the
ensureGeneralInstructions helper out of FleetManager into a standalone
module. Logger is now an optional argument so the function stays free
of FleetManager state.

Pure data + pure helper extraction — no behaviour change, no new deps.

P4.1 step 2 of 4.
